Add a check in graph assembly to fail assembly if we can't identify a reference source or sink.
This is most likely to happen if we have a long repetitive region that forms a cycle in the reference strand.
This prevents us from calling variants in this active region.  Previously, GATK simply crashed on this error, now we fail
gracefully and move on to the next active region.